+++ .github/workflows/storybook_deploy.yml
@@ -1,0 +1,28 @@
+name: Build & Deploy Storybook
+on:
+  pull_request:
+    branches: [dev]
+    paths: ['src/components/common/**']
+
+jobs:
+  test:
+    permissions:
+      pull-requests: write
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v3
+      - name : Install dependencies
+        run : npm install
+      - name : Publish to Chromatic
+        id: chromatic
+        uses: chromaui/action@v1
+        with :
+          projectToken: ${{ secrets.CHROMATIC_PROJECT_TOKEN }}
+          token: ${{secrets.GITHUB_TOKEN}}
+      - name: comment PR  
+        uses: thollander/actions-comment-pull-request@v2  
+        env:
+          GITHUB_TOEKN: ${{secrets.ACTION_TOKEN}}
+        with:  
+          message: "🚀storybook: ${{steps.chromatic.outputs.stroybookUrl}}"
